Simple sender and receiver
Write a simple sender and receiver feature. When the program is launched without arguments, it waits for a connection to receive file. Providing the "send" argument instead, and specifying a destination and the path for a file to send, the file is sent to another instance of the same program.

Signed-off-by: Alessio Pascucci <feeder81@gmail.com>